requirements.txt
pyTelegramBotAPI
google-generativeai
Pillow
requests
fastapi
uvicorn
gunicorn
python-multipart